requirements.txt
flask
uvicorn[standard]
gunicorn
psycopg2.binary